infra/ctfd-oauth-plugin/requirements.txt
flask-oauthlib
